Making Ids long and modify repositories

+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Commands/AcceptInvitation/AcceptInvitationCommandHandler.cs
@@ -56,7 +56,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Application.Commands.AcceptInvi
             if (existing.ExpiryDate < DateTimeProvider.Current.UtcNow)
                 throw new InvalidRequestException(new Dictionary<string, string> { { "Invitation", "Invitation has expired" } });
 
-            await _invitationRepository.Accept(existing.Email, existing.AccountId, (int) existing.RoleId);
+            await _invitationRepository.Accept(existing.Email, existing.AccountId, (short) existing.RoleId);
         }
     }
 }
