+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Commands/ChangeTeamMemberRole/ChangeTeamMemberRoleCommand.cs
@@ -6,7 +6,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Application.Commands.ChangeTeam
     {
         public long AccountId { get; set; }
         public string Email { get; set; }
-        public int RoleId { get; set; }
+        public short RoleId { get; set; }
         public string ExternalUserId { get; set; }
     }
 }
