+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Data/IAccountTeamRepository.cs
@@ -5,7 +5,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Domain.Data
 {
     public interface IAccountTeamRepository
     {
-        Task<List<TeamMember>> GetAccountTeamMembersForUserId(int accountId, string externalUserId);
+        Task<List<TeamMember>> GetAccountTeamMembersForUserId(long accountId, string externalUserId);
         Task<TeamMember> GetMember(long accountId, string email);
     }
 }
