+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Data/IAggregationRepository.cs
@@ -4,8 +4,8 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Domain.Data
 {
     public interface IAggregationRepository
     {
-        Task Update(int accountId, int pageNumber, string json);
+        Task Update(long accountId, int pageNumber, string json);
 
-        Task<AggregationData> GetByAccountId(int accountId);
+        Task<AggregationData> GetByAccountId(long accountId);
     }
 }
