+++ src/SFA.DAS.EmployerApprenticeshipsService.Infrastructure/Data/InvitationRepository.cs
@@ -38,7 +38,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Infrastructure.Data
             var result = await WithConnection(async c =>
             {
                 var parameters = new DynamicParameters();
-                parameters.Add("@id", id, DbType.Int32);
+                parameters.Add("@id", id, DbType.Int64);
 
                 return await c.QueryAsync<InvitationView>(
                     sql: "SELECT * FROM [dbo].[GetInvitations] WHERE Id = @id;",
@@ -54,12 +54,12 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Infrastructure.Data
             await WithConnection(async c =>
             {
                 var parameters = new DynamicParameters();
-                parameters.Add("@accountId", invitation.AccountId, DbType.Int32);
+                parameters.Add("@accountId", invitation.AccountId, DbType.Int64);
                 parameters.Add("@name", invitation.Name, DbType.String);
                 parameters.Add("@email", invitation.Email, DbType.String);
                 parameters.Add("@expiryDate", invitation.ExpiryDate, DbType.DateTime);
                 parameters.Add("@statusId", invitation.Status, DbType.Int16);
-                parameters.Add("@roleId", invitation.RoleId, DbType.Int32);
+                parameters.Add("@roleId", invitation.RoleId, DbType.Int16);
 
                 return await c.ExecuteAsync(
                     sql: "INSERT INTO [dbo].[Invitation] ([AccountId],[Name],[Email],[ExpiryDate],[Status],[RoleId]) VALUES (@accountId, @name, @email, @expiryDate, @statusId, @roleId)",
@@ -73,7 +73,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Infrastructure.Data
             var result = await WithConnection(async c =>
             {
                 var parameters = new DynamicParameters();
-                parameters.Add("@id", id, DbType.Int32);
+                parameters.Add("@id", id, DbType.Int64);
 
                 return await c.QueryAsync<Invitation>(
                     sql: "SELECT * FROM [dbo].[Invitation] WHERE Id = @id;",
@@ -89,7 +89,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Infrastructure.Data
             var result = await WithConnection(async c =>
             {
                 var parameters = new DynamicParameters();
-                parameters.Add("@accountId", accountId, DbType.Int32);
+                parameters.Add("@accountId", accountId, DbType.Int64);
                 parameters.Add("@email", email, DbType.String);
 
                 return await c.QueryAsync<Invitation>(
@@ -106,7 +106,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Infrastructure.Data
             await WithConnection(async c =>
             {
                 var parameters = new DynamicParameters();
-                parameters.Add("@id", invitation.Id, DbType.Int32);
+                parameters.Add("@id", invitation.Id, DbType.Int64);
                 parameters.Add("@statusId", invitation.Status, DbType.Int16);
 
                 return await c.ExecuteAsync(
@@ -121,7 +121,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Infrastructure.Data
             await WithConnection(async c =>
             {
                 var parameters = new DynamicParameters();
-                parameters.Add("@id", invitation.Id, DbType.Int32);
+                parameters.Add("@id", invitation.Id, DbType.Int64);
                 parameters.Add("@name", invitation.Name, DbType.String);
                 parameters.Add("@roleId", invitation.RoleId, DbType.Int16);
                 parameters.Add("@statusId", invitation.Status, DbType.Int16);
@@ -134,13 +134,13 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Infrastructure.Data
             });
         }
 
-        public async Task Accept(string email, long accountId, int roleId)
+        public async Task Accept(string email, long accountId, short roleId)
         {
             await WithConnection(async c =>
             {
                 var parameters = new DynamicParameters();
                 parameters.Add("@email", email, DbType.String);
-                parameters.Add("@accountId", accountId, DbType.Int32);
+                parameters.Add("@accountId", accountId, DbType.Int64);
                 parameters.Add("@roleId", roleId, DbType.Int16);
 
                 return await c.ExecuteAsync(
